requirements.txt
streamlit
fastapi
uvicorn
mediapipe
numpy
opencv-python-headless
requests
python-multipart
pillow